refactor(terminal-output): split 2.5k-line monolith into folder module
Move TerminalOutput into TerminalOutput/ with utils, hooks, and subcomponents
so the root orchestrator drops to ~475 lines. Preserve behavior and the
existing ../TerminalOutput import path via an index.ts barrel.
