--- src/renderer/components/TerminalOutput.tsx
@@ -1,2566 +1,0 @@
-import React, { useRef, useEffect, useMemo, forwardRef, useState, useCallback, memo } from 'react';
-import {
-	ChevronDown,
-	ChevronUp,
-	Trash2,
-	Copy,
-	Check,
-	ArrowDown,
-	Eye,
-	FileText,
-	RotateCcw,
-	AlertCircle,
-	Save,
-	Share2,
-	Hammer,
-	GitFork,
-} from 'lucide-react';
-import type { Session, Theme, LogEntry, FocusArea, AgentError, QueuedItem } from '../types';
-import type { FileNode } from '../types/fileTree';
-import Convert from 'ansi-to-html';
-import { useLayerStack } from '../contexts/LayerStackContext';
-import { MODAL_PRIORITIES } from '../constants/modalPriorities';
-import { getActiveTab } from '../utils/tabHelpers';
-import { useDebouncedValue, useThrottledCallback } from '../hooks';
-import {
-	processLogTextHelper,
-	filterTextByLinesHelper,
-	getCachedAnsiHtml,
-} from '../utils/textProcessing';
-import { jumpToMessageEdge, isTextInputTarget } from '../utils/messageScrollNavigation';
-import { JumpToMessageTopButton } from './JumpToMessageTopButton';
-import { formatShortcutKeys } from '../utils/shortcutFormatter';
-import { MarkdownRenderer } from './MarkdownRenderer';
-import { QueuedItemsList } from './QueuedItemsList';
-import { LogFilterControls } from './LogFilterControls';
-import { SaveMarkdownModal } from './SaveMarkdownModal';
-import { generateTerminalProseStyles } from '../utils/markdownConfig';
-import { linkifyNode } from '../utils/linkify';
-import { safeClipboardWrite } from '../utils/clipboard';
-import { flashCopiedToClipboard } from '../utils/flashCopiedToClipboard';
-import { useSettingsStore } from '../stores/settingsStore';
-import { useMessageGistStore } from '../stores/messageGistStore';
-import { useSessionStore } from '../stores/sessionStore';
-import { SessionRecoveryCard } from './SessionRecoveryCard';
-import { RetryStatusCard } from './RetryStatusCard';
-import { getTokenSourcePill } from '../../shared/claudeTokenModeLabel';
-import { getClaudeTokenMode } from '../../shared/claudeTokenMode';
-import { CrossAgentResponseHeader } from './CrossAgentResponseHeader';
-
-// ============================================================================
-// Tool display helpers (pure functions, hoisted out of render path)
-// ============================================================================
-
-/**
- * Collapse a tab's logs into render groups for the AI transcript.
- *
- * Consecutive LOCAL response entries (the active agent's own stdout/ai/system
- * output for one turn) fold into ONE bubble per user message, so a turn made of
- * several streamed entries reads as a single reply. These kinds stay standalone:
- *   - `user` messages,
- *   - `tool` / `thinking` entries,
- *   - Agent Resilience outage markers (`retryOutageId`), which render as a live
- *     status card and must not fold into a text group,
- *   - cross-agent (`@mention`) replies. Each cross-agent reply already streams
- *     into its own single entry and carries its own attribution header, so it
- *     must never fold into the local response group OR into a sibling
- *     cross-agent reply. Without this, several agents' answers (plus the local
- *     one) merge into a single bubble that inherits the FIRST entry's provenance
- *     - e.g. two remote replies and the local one all rendering inside one
- *     agent's error bubble.
- *
- * Pure + exported so the grouping is unit-testable independent of the component.
- */
-export function collapseAiResponseLogs(logs: LogEntry[]): LogEntry[] {
-	const result: LogEntry[] = [];
-	let currentResponseGroup: LogEntry[] = [];
-
-	// Flush the accumulated LOCAL response group into one combined entry.
-	const flushResponseGroup = () => {
-		if (currentResponseGroup.length > 0) {
-			const combinedText = currentResponseGroup.map((l) => l.text).join('');
-			// The token-source pill keys off `renderStyle === 'text-stream'`
-			// (maestro-p TUI capture). A response group can lead with a non-stream
-			// entry (e.g. the "Adaptive Mode: switched ..." system banner), and
-			// basing the combined entry only on `[0]` would inherit that entry's
-			// missing renderStyle and mislabel an interactive turn as "API".
-			// Preserve text-stream if ANY grouped entry carries it.
-			const hasTextStream = currentResponseGroup.some((l) => l.renderStyle === 'text-stream');
-			result.push({
-				...currentResponseGroup[0],
-				text: combinedText,
-				// Keep the first entry's timestamp and id.
-				...(hasTextStream ? { renderStyle: 'text-stream' as const } : {}),
-			});
-			currentResponseGroup = [];
-		}
-	};
-
-	for (const log of logs) {
-		if (log.source === 'user') {
-			flushResponseGroup();
-			result.push(log);
-		} else if (log.source === 'tool' || log.source === 'thinking' || log.retryOutageId) {
-			// Flush the response group, then keep tool/thinking and Agent Resilience
-			// outage markers as their own entries. The outage marker must not merge
-			// into a text group — it renders as a live status card.
-			flushResponseGroup();
-			result.push(log);
-		} else if (log.metadata?.crossAgent) {
-			// Standalone, individually-attributed bubble (see fn doc).
-			flushResponseGroup();
-			result.push(log);
-		} else {
-			// Accumulate the local agent's own response entries.
-			currentResponseGroup.push(log);
-		}
-	}
-
-	flushResponseGroup();
-	return result;
-}
-
-/** Handle command values that may be strings or string arrays (Codex uses arrays) */
-const safeCommand = (v: unknown): string | null => {
-	if (typeof v === 'string') return v;
-	if (Array.isArray(v) && v.length > 0 && v.every((x) => typeof x === 'string')) {
-		return v.join(' ');
-	}
-	return null;
-};
-
-/** Summarize TodoWrite todos array — shows in-progress task and progress count */
-const summarizeTodos = (v: unknown): string | null => {
-	if (!Array.isArray(v) || v.length === 0) return null;
-	const todos = v as Array<{ content?: string; status?: string; activeForm?: string }>;
-	const completed = todos.filter((t) => t.status === 'completed').length;
-	const inProgress = todos.find((t) => t.status === 'in_progress');
-	const label = inProgress?.activeForm || inProgress?.content || todos[0]?.content;
-	if (!label) return `${todos.length} tasks`;
-	return `${label} (${completed}/${todos.length})`;
-};
-
-/** Structured result from summarizeToolInput for richer rendering */
-interface ToolSummary {
-	/** Human-readable description (e.g. Bash description field) */
-	description?: string;
-	/** Primary content — command text or generic summary */
-	detail: string;
-}
-
-/**
- * Summarize tool input generically — no per-tool extractors needed.
- * Returns structured data so the renderer can display description and command
- * with proper visual hierarchy.
- *
- * Tool logs are only emitted when thinking is enabled, so we show the full
- * command text without truncation to give complete visibility into agent actions.
- */
-const summarizeToolInput = (input: unknown): ToolSummary | null => {
-	// Some agents (notably Copilot/Codex apply_patch) deliver the tool argument
-	// as a raw string instead of an object — Object.entries on a string would
-	// iterate it character-by-character and produce garbled, space-separated
-	// output, so surface the string as-is.
-	if (typeof input === 'string') {
-		return input ? { detail: input } : null;
-	}
-	if (!input || typeof input !== 'object' || Array.isArray(input)) {
-		return null;
-	}
-	const inputRecord = input as Record<string, unknown>;
-
-	// Special case: TodoWrite todos array
-	const todosResult = summarizeTodos(inputRecord.todos);
-	if (todosResult) return { detail: todosResult };
-
-	// Extract description field separately for structured display
-	const description =
-		typeof inputRecord.description === 'string' && inputRecord.description
-			? inputRecord.description
-			: undefined;
-
-	// Collect displayable values (skip huge blobs)
-	const parts: string[] = [];
-	for (const [key, val] of Object.entries(inputRecord)) {
-		if (val === undefined || val === null || val === '') continue;
-		// Skip description — rendered separately
-		if (key === 'description') continue;
-		// Command arrays (Codex)
-		const cmd = safeCommand(val);
-		if (cmd) {
-			parts.push(cmd);
-			continue;
-		}
-		// Arrays: show count
-		if (Array.isArray(val)) {
-			parts.push(`${key}: [${val.length}]`);
-			continue;
-		}
-		// Objects: skip (too noisy)
-		if (typeof val === 'object') continue;
-		// Booleans/numbers: show as key=value
-		if (typeof val === 'boolean' || typeof val === 'number') {
-			parts.push(`${key}=${val}`);
-			continue;
-		}
-	}
-	const detail = parts.length > 0 ? parts.join('  ') : undefined;
-	if (!detail && !description) return null;
-	return { description, detail: detail ?? '' };
-};
-
-/** Max lines of tool output to preview inline before truncating. */
-const TOOL_OUTPUT_PREVIEW_LINES = 8;
-
-/**
- * Summarize tool output for inline display. MCP tools (and others) return their
- * result in `toolState.output`; without this the compact tool log shows only the
- * name + status icon and drops the result entirely. Strings render as-is, objects
- * are JSON-stringified. Output is capped to a short preview so large results don't
- * flood the chat.
- */
-const summarizeToolOutput = (output: unknown): string | null => {
-	if (output === undefined || output === null) return null;
-	let text: string;
-	if (typeof output === 'string') {
-		text = output;
-	} else {
-		try {
-			text = JSON.stringify(output, null, 2);
-		} catch {
-			return null;
-		}
-	}
-	text = text.trim();
-	if (!text) return null;
-	const lines = text.split('\n');
-	if (lines.length > TOOL_OUTPUT_PREVIEW_LINES) {
-		return lines.slice(0, TOOL_OUTPUT_PREVIEW_LINES).join('\n') + '\n…';
-	}
-	return text;
-};
-
-const isHiddenProgressEntry = (log: LogEntry): boolean =>
-	log.source === 'system' && log.id.startsWith('hidden-progress:');
-
-/**
- * Connector that reads the live tab from the session store and renders the
- * SessionRecoveryCard. Keeps LogItem's prop surface narrow (just sessionId)
- * instead of passing the full Session object through every log entry.
- */
-function SessionRecoveryCardConnector(props: {
-	theme: Theme;
-	sessionId: string;
-	recoveryAction: NonNullable<LogEntry['recoveryAction']>;
-	isRecovering: boolean;
-	recoveryError: string | null;
-	onRecover: (opts: {
-		sessionId: string;
-		tabId: string;
-		lastUserPrompt: string;
-		groomContext: boolean;
-	}) => void;
-}) {
-	const tab = useSessionStore((s) => {
-		const session = s.sessions.find((sess) => sess.id === props.sessionId);
-		return session?.aiTabs.find((t) => t.id === props.recoveryAction.tabId);
-	});
-	if (!tab) return null;
-	return (
-		<SessionRecoveryCard
-			theme={props.theme}
-			sessionId={props.sessionId}
-			tab={tab}
-			lastUserPrompt={props.recoveryAction.lastUserPrompt}
-			isRecovering={props.isRecovering}
-			recoveryError={props.recoveryError}
-			onRecover={props.onRecover}
-		/>
-	);
-}
-
-// ============================================================================
-// LogItem - Memoized component for individual log entries
-// ============================================================================
-
-interface LogItemProps {
-	log: LogEntry;
-	index: number;
-	isTerminal: boolean;
-	isAIMode: boolean;
-	theme: Theme;
-	fontFamily: string;
-	maxOutputLines: number;
-	lastUserCommand?: string;
-	// Expansion state
-	isExpanded: boolean;
-	onToggleExpanded: (logId: string) => void;
-	// Local filter state
-	localFilterQuery: string;
-	filterMode: { mode: 'include' | 'exclude'; regex: boolean };
-	activeLocalFilter: string | null;
-	onToggleLocalFilter: (logId: string) => void;
-	onSetLocalFilterQuery: (logId: string, query: string) => void;
-	onSetFilterMode: (
-		logId: string,
-		update: (current: { mode: 'include' | 'exclude'; regex: boolean }) => {
-			mode: 'include' | 'exclude';
-			regex: boolean;
-		}
-	) => void;
-	onClearLocalFilter: (logId: string) => void;
-	// Delete state
-	deleteConfirmLogId: string | null;
-	onDeleteLog?: (logId: string) => number | null;
-	onSetDeleteConfirmLogId: (logId: string | null) => void;
-	scrollContainerRef: React.RefObject<HTMLDivElement>;
-	// Other callbacks
-	setLightboxImage: (
-		image: string | null,
-		contextImages?: string[],
-		source?: 'staged' | 'history'
-	) => void;
-	copyToClipboard: (text: string) => void;
-	// ANSI converter
-	ansiConverter: Convert;
-	// Markdown rendering mode for AI responses (when true, shows raw text)
-	markdownEditMode: boolean;
-	onToggleMarkdownEditMode: () => void;
-	// Replay message callback (AI mode only)
-	onReplayMessage?: (text: string, images?: string[]) => void;
-	// File linking support
-	fileTree?: FileNode[];
-	cwd?: string;
-	projectRoot?: string;
-	onFileClick?: (path: string) => void;
-	// SSH remote ID for resolving image/file paths emitted by an agent running
-	// on a remote host. Without it, LocalImage reads the path from the local
-	// filesystem (which doesn't have the remote file) and the image fails.
-	sshRemoteId?: string;
-	// Error details callback - receives the specific AgentError from the log entry
-	onShowErrorDetails?: (error: AgentError) => void;
-	// Save to file callback (AI mode only, non-user messages)
-	onSaveToFile?: (text: string) => void;
-	// Publish to GitHub Gist (AI mode only, non-user messages, requires gh CLI)
-	ghCliAvailable?: boolean;
-	onPublishGist?: (text: string, messageId?: string) => void;
-	publishedGistUrl?: string;
-	// Fork conversation from this message (AI mode only, user messages and AI responses — source 'user' | 'ai' | 'stdout')
-	onForkConversation?: (logId: string) => void;
-	bionifyReadingMode: boolean;
-	bionifyIntensity: number;
-	bionifyAlgorithm: string;
-	// Message alignment
-	userMessageAlignment: 'left' | 'right';
-	// Claude mode pill — both passed as primitives so LogItem memo equality stays cheap.
-	isClaudeCode: boolean;
-	isAdaptiveMode: boolean;
-	// Session recovery (session_not_found inline card). Only consumed when
-	// log.recoveryAction is set; otherwise these props are ignored.
-	sessionId: string;
-	onSessionRecover?: (opts: {
-		sessionId: string;
-		tabId: string;
-		lastUserPrompt: string;
-		groomContext: boolean;
-	}) => void;
-	isRecoveringSession?: boolean;
-	sessionRecoveryError?: string | null;
-}
-
-const LogItemComponent = memo(
-	({
-		log,
-		index,
-		isTerminal,
-		isAIMode,
-		theme,
-		fontFamily,
-		maxOutputLines,
-		lastUserCommand,
-		isExpanded,
-		onToggleExpanded,
-		localFilterQuery,
-		filterMode,
-		activeLocalFilter,
-		onToggleLocalFilter,
-		onSetLocalFilterQuery,
-		onSetFilterMode,
-		onClearLocalFilter,
-		deleteConfirmLogId,
-		onDeleteLog,
-		onSetDeleteConfirmLogId,
-		scrollContainerRef,
-		setLightboxImage,
-		copyToClipboard,
-		ansiConverter,
-		markdownEditMode,
-		onToggleMarkdownEditMode,
-		onReplayMessage,
-		fileTree,
-		cwd,
-		projectRoot,
-		onFileClick,
-		sshRemoteId,
-		onShowErrorDetails,
-		onSaveToFile,
-		ghCliAvailable,
-		onPublishGist,
-		publishedGistUrl,
-		onForkConversation,
-		bionifyReadingMode,
-		bionifyIntensity,
-		bionifyAlgorithm,
-		userMessageAlignment,
-		isClaudeCode,
-		isAdaptiveMode,
-		sessionId,
-		onSessionRecover,
-		isRecoveringSession,
-		sessionRecoveryError,
-	}: LogItemProps) => {
-		// Ref for the log item container - used for scroll-into-view on expand
-		const logItemRef = useRef<HTMLDivElement>(null);
-
-		// Handle expand toggle with scroll adjustment
-		const handleExpandToggle = useCallback(() => {
-			const wasExpanded = isExpanded;
-			onToggleExpanded(log.id);
-
-			// After expanding, scroll to ensure the bottom of the item is visible
-			if (!wasExpanded) {
-				// Use setTimeout to wait for the DOM to update after expansion
-				setTimeout(() => {
-					const logItem = logItemRef.current;
-					const container = scrollContainerRef.current;
-					if (logItem && container) {
-						const itemRect = logItem.getBoundingClientRect();
-						const containerRect = container.getBoundingClientRect();
-
-						// Check if the bottom of the item is below the visible area
-						const itemBottom = itemRect.bottom;
-						const containerBottom = containerRect.bottom;
-
-						if (itemBottom > containerBottom) {
-							// Scroll to show the bottom of the item with some padding
-							const scrollAmount = itemBottom - containerBottom + 20; // 20px padding
-							container.scrollBy({ top: scrollAmount, behavior: 'smooth' });
-						}
-					}
-				}, 50); // Small delay to allow React to re-render
-			}
-		}, [isExpanded, log.id, onToggleExpanded, scrollContainerRef]);
-
-		// Strip command echo from terminal output
-		let textToProcess = log.text;
-		if (isTerminal && log.source !== 'user' && lastUserCommand) {
-			if (textToProcess.startsWith(lastUserCommand)) {
-				textToProcess = textToProcess.slice(lastUserCommand.length);
-				if (textToProcess.startsWith('\r\n')) {
-					textToProcess = textToProcess.slice(2);
-				} else if (textToProcess.startsWith('\n') || textToProcess.startsWith('\r')) {
-					textToProcess = textToProcess.slice(1);
-				}
-			}
-		}
-
-		const processedText = processLogTextHelper(textToProcess, isTerminal && log.source !== 'user');
-
-		// Skip rendering stderr entries that have no actual content
-		if (log.source === 'stderr' && !processedText.trim()) {
-			return null;
-		}
-
-		// Separate stdout and stderr for terminal output
-		const separated =
-			log.source === 'stderr'
-				? { stdout: '', stderr: processedText }
-				: { stdout: processedText, stderr: '' };
-
-		// Apply local filter if active for this log entry
-		const filteredStdout =
-			localFilterQuery && log.source !== 'user'
-				? filterTextByLinesHelper(
-						separated.stdout,
-						localFilterQuery,
-						filterMode.mode,
-						filterMode.regex
-					)
-				: separated.stdout;
-		const filteredStderr =
-			localFilterQuery && log.source !== 'user'
-				? filterTextByLinesHelper(
-						separated.stderr,
-						localFilterQuery,
-						filterMode.mode,
-						filterMode.regex
-					)
-				: separated.stderr;
-
-		// Check if filter returned no results
-		const hasNoMatches =
-			localFilterQuery && !filteredStdout.trim() && !filteredStderr.trim() && log.source !== 'user';
-
-		// For stderr entries, use stderr content; for all others, use stdout content
-		const contentToDisplay = log.source === 'stderr' ? filteredStderr : filteredStdout;
-
-		// PERF: Convert ANSI codes to HTML using cache.
-		// Search highlighting is now applied at the scroll-container level via CSS Custom
-		// Highlight API in TerminalOutput, so per-log markers are no longer needed.
-		const htmlContent =
-			isTerminal && log.source !== 'user'
-				? getCachedAnsiHtml(contentToDisplay, theme.id, ansiConverter)
-				: contentToDisplay;
-
-		const filteredText = contentToDisplay;
-
-		// Count lines in the filtered text
-		const lineCount = filteredText.split('\n').length;
-		const shouldCollapse = lineCount > maxOutputLines && maxOutputLines !== Infinity;
-
-		// Truncate text if collapsed
-		const displayText =
-			shouldCollapse && !isExpanded
-				? filteredText.split('\n').slice(0, maxOutputLines).join('\n')
-				: filteredText;
-
-		// PERF: Sanitize with DOMPurify, using cache for ANSI conversion.
-		// Search highlighting is handled at the scroll-container level.
-		const displayHtmlContent =
-			shouldCollapse && !isExpanded && isTerminal && log.source !== 'user'
-				? getCachedAnsiHtml(displayText, theme.id, ansiConverter)
-				: htmlContent;
-
-		const isUserMessage = log.source === 'user';
-		const isReversed = isUserMessage
-			? userMessageAlignment === 'left'
-			: userMessageAlignment === 'right';
-
-		// Agent Resilience: an outage marker renders as a live status card in a
-		// clean row (no error-tinted bubble chrome), left gutter kept for alignment.
-		if (log.retryOutageId) {
-			return (
-				<div
-					ref={logItemRef}
-					className="flex gap-4 px-3 sm:px-6 py-2"
-					data-log-index={index}
-					style={{ contentVisibility: 'auto', containIntrinsicSize: 'auto 120px' }}
-				>
-					<div className="hidden sm:block w-20 shrink-0" />
-					<div className="flex-1 min-w-0">
-						<RetryStatusCard outageId={log.retryOutageId} theme={theme} fallbackText={log.text} />
-					</div>
-				</div>
-			);
-		}
-
-		// Cross-agent (@@mention) reply provenance. When set, this AI entry was
-		// produced by a DIFFERENT agent the user consulted; it gets a tinted
-		// bubble + attribution pill (Phase 04). Themes may override the tint via
-		// the crossAgentBubbleBg / crossAgentBubbleBorder tokens; otherwise we
-		// derive a subtle accent wash with the same color-mix idiom the user
-		// bubble uses below, so every theme reads correctly for free.
-		const crossAgent = log.metadata?.crossAgent;
-		const isCrossAgentStreaming = !!crossAgent?.streaming;
-		// Phase 05: a failed consult surfaces inline as a red-tinted variant of the
-		// cross-agent bubble (never throws). Otherwise the normal accent wash.
-		const isCrossAgentError = !!crossAgent?.error;
-		const crossAgentBubbleBg = isCrossAgentError
-			? `color-mix(in srgb, ${theme.colors.error} 12%, ${theme.colors.bgActivity})`
-			: (theme.colors.crossAgentBubbleBg ??
-				`color-mix(in srgb, ${theme.colors.accent} 14%, ${theme.colors.bgActivity})`);
-		const crossAgentBubbleBorder = isCrossAgentError
-			? theme.colors.error + '66'
-			: (theme.colors.crossAgentBubbleBorder ?? theme.colors.accent + '55');
-
-		return (
-			<div
-				ref={logItemRef}
-				// Narrow screens (phones / web-desktop mobile): the fixed side gutter
-				// for the timestamp costs ~96px of bubble width, so the row stacks -
-				// timestamp above, bubble full-width. From `sm` up it's the classic
-				// side-by-side layout with the w-20 timestamp column.
-				className={`flex flex-col gap-1 sm:gap-4 group ${isReversed ? 'sm:flex-row-reverse' : 'sm:flex-row'} px-3 sm:px-6 py-2`}
-				data-log-index={index}
-				// PERF: the transcript is not virtualized, so every message stays in the
-				// DOM. content-visibility:auto lets the browser skip style/layout/paint for
-				// off-screen rows (the dominant scroll cost - a huge static layer tree the
-				// compositor re-walked every frame). contain-intrinsic-size: 'auto <fallback>'
-				// reserves height so the scrollbar stays stable; `auto` makes the browser
-				// remember each row's real rendered size after it's been shown once. No DOM,
-				// React, or behavior change - purely a rendering hint.
-				style={{ contentVisibility: 'auto', containIntrinsicSize: 'auto 120px' }}
-			>
-				<div
-					className={`shrink-0 text-[10px] sm:w-20 sm:pt-2 flex gap-1 sm:block ${isReversed ? 'text-right justify-end' : 'text-left'}`}
-					style={{ fontFamily, color: theme.colors.textDim, opacity: 0.6 }}
-				>
-					{(() => {
-						const logDate = new Date(log.timestamp);
-						const today = new Date();
-						const isToday = logDate.toDateString() === today.toDateString();
-						const time = logDate.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
-						if (isToday) {
-							return time;
-						}
-						// Format: YYYY-MM-DD on first line, time on second
-						const year = logDate.getFullYear();
-						const month = String(logDate.getMonth() + 1).padStart(2, '0');
-						const day = String(logDate.getDate()).padStart(2, '0');
-						return (
-							<>
-								<div>
-									{year}-{month}-{day}
-								</div>
-								<div>{time}</div>
-							</>
-						);
-					})()}
-				</div>
-				<div
-					className={`flex-1 min-w-0 p-4 pb-10 rounded-xl border ${isReversed ? 'rounded-tr-none' : 'rounded-tl-none'} relative overflow-hidden ${isCrossAgentStreaming ? 'animate-status-glow' : ''}`}
-					style={{
-						backgroundColor: crossAgent
-							? crossAgentBubbleBg
-							: isUserMessage
-								? isAIMode
-									? `color-mix(in srgb, ${theme.colors.accent} 20%, ${theme.colors.bgSidebar})`
-									: `color-mix(in srgb, ${theme.colors.accent} 15%, ${theme.colors.bgActivity})`
-								: log.source === 'stderr' || log.source === 'error'
-									? `color-mix(in srgb, ${theme.colors.error} 8%, ${theme.colors.bgActivity})`
-									: isAIMode
-										? theme.colors.bgActivity
-										: 'transparent',
-						borderColor: crossAgent
-							? crossAgentBubbleBorder
-							: isUserMessage && isAIMode
-								? theme.colors.accent + '40'
-								: log.source === 'stderr' || log.source === 'error'
-									? theme.colors.error
-									: theme.colors.border,
-						// Drives the accent hue of the streaming pulse (.animate-status-glow).
-						...(isCrossAgentStreaming
-							? ({ '--status-glow-color': theme.colors.accent } as React.CSSProperties)
-							: {}),
-					}}
-				>
-					{/* Cross-agent attribution header - this AI reply was routed back from a
-					    DIFFERENT agent the user consulted via @mention. Names the agent,
-					    provider, and session id, and offers two ways to jump into that
-					    agent to continue the dialogue. Streaming shows a spinner; a failed
-					    consult tints it red. In a group fan-out it's what tells each
-					    response apart. */}
-					{crossAgent && <CrossAgentResponseHeader crossAgent={crossAgent} theme={theme} />}
-					{/* Local filter icon for system output only */}
-					{log.source !== 'user' && isTerminal && (
-						<div className="absolute top-2 right-2 flex items-center gap-2">
-							<LogFilterControls
-								logId={log.id}
-								fontFamily={fontFamily}
-								theme={theme}
-								filterQuery={localFilterQuery}
-								filterMode={filterMode}
-								isActive={activeLocalFilter === log.id}
-								onToggleFilter={onToggleLocalFilter}
-								onSetFilterQuery={onSetLocalFilterQuery}
-								onSetFilterMode={onSetFilterMode}
-								onClearFilter={onClearLocalFilter}
-							/>
-						</div>
-					)}
-					{log.images && log.images.length > 0 && (
-						<div
-							className="flex gap-2 mb-2 overflow-x-auto scrollbar-thin"
-							style={{ overscrollBehavior: 'contain' }}
-						>
-							{log.images.map((img, imgIdx) => (
-								<button
-									key={`${img}-${imgIdx}`}
-									type="button"
-									className="shrink-0 p-0 bg-transparent outline-none focus:ring-2 focus:ring-accent rounded"
-									onClick={() => setLightboxImage(img, log.images, 'history')}
-								>
-									<img
-										src={img}
-										alt={`Terminal output image ${imgIdx + 1}`}
-										className="h-20 rounded border cursor-zoom-in block"
-										style={{ objectFit: 'contain', maxWidth: '200px' }}
-									/>
-								</button>
-							))}
-						</div>
-					)}
-					{log.source === 'stderr' && (
-						<div className="mb-2">
-							<span
-								className="px-2 py-1 rounded text-xs font-bold uppercase tracking-wide"
-								style={{
-									backgroundColor: theme.colors.error,
-									color: '#fff',
-								}}
-							>
-								STDERR
-							</span>
-						</div>
-					)}
-					{/* Special rendering for error log entries */}
-					{log.source === 'error' && (
-						<div className="flex flex-col gap-3">
-							<div className="flex items-center gap-2">
-								<AlertCircle className="w-5 h-5" style={{ color: theme.colors.error }} />
-								<span className="text-sm font-medium" style={{ color: theme.colors.error }}>
-									Error
-								</span>
-							</div>
-							<div className="text-sm" style={{ color: theme.colors.textMain }}>
-								<MarkdownRenderer
-									content={log.text}
-									theme={theme}
-									onCopy={copyToClipboard}
-									fileTree={fileTree}
-									cwd={cwd}
-									projectRoot={projectRoot}
-									onFileClick={onFileClick}
-									sshRemoteId={sshRemoteId}
-									chatLineBreaks
-									chatMath
-								/>
-							</div>
-							{!!log.agentError?.parsedJson && onShowErrorDetails && (
-								<button
-									onClick={() => onShowErrorDetails(log.agentError!)}
-									className="self-start flex items-center gap-2 px-3 py-1.5 text-xs rounded border hover:opacity-80 transition-opacity"
-									style={{
-										backgroundColor: theme.colors.error + '15',
-										borderColor: theme.colors.error + '40',
-										color: theme.colors.error,
-									}}
-								>
-									<Eye className="w-3 h-3" />
-									View Details
-								</button>
-							)}
-						</div>
-					)}
-					{/* Special rendering for thinking/streaming content (AI reasoning in real-time) */}
-					{log.source === 'thinking' && (
-						<div
-							className="px-4 py-2 text-sm font-mono border-l-2"
-							style={{
-								color: theme.colors.textMain,
-								borderColor: theme.colors.accent,
-							}}
-						>
-							<div className="flex items-center gap-2 mb-1">
-								<span
-									className="text-[10px] px-1.5 py-0.5 rounded"
-									style={{
-										backgroundColor: `${theme.colors.accent}30`,
-										color: theme.colors.accent,
-									}}
-								>
-									thinking
-								</span>
-							</div>
-							<div className="whitespace-pre-wrap text-sm break-words">
-								{isAIMode && !markdownEditMode ? (
-									<MarkdownRenderer
-										content={log.text}
-										theme={theme}
-										onCopy={copyToClipboard}
-										enableBionifyReadingMode={bionifyReadingMode}
-										bionifyIntensity={bionifyIntensity}
-										bionifyAlgorithm={bionifyAlgorithm}
-										fileTree={fileTree}
-										cwd={cwd}
-										projectRoot={projectRoot}
-										onFileClick={onFileClick}
-										sshRemoteId={sshRemoteId}
-										chatLineBreaks
-										chatMath
-									/>
-								) : (
-									log.text
-								)}
-							</div>
-						</div>
-					)}
-					{isHiddenProgressEntry(log) && (
-						<div
-							className="px-4 py-1.5 text-xs border-l-2"
-							style={{
-								color: theme.colors.textMain,
-								borderColor: theme.colors.accent,
-							}}
-						>
-							<div className="flex items-start gap-2">
-								<span
-									className="px-1.5 py-0.5 rounded shrink-0"
-									style={{
-										backgroundColor: `${theme.colors.accent}30`,
-										color: theme.colors.accent,
-									}}
-								>
-									{log.metadata?.hiddenProgress?.kind === 'tool'
-										? log.metadata.hiddenProgress.toolName || 'working'
-										: 'thinking'}
-								</span>
-								{log.metadata?.toolState?.status === 'completed' ? (
-									<span className="shrink-0 pt-0.5" style={{ color: theme.colors.success }}>
-										✓
-									</span>
-								) : log.metadata?.toolState?.status === 'failed' ||
-								  log.metadata?.toolState?.status === 'error' ? (
-									<span className="shrink-0 pt-0.5" style={{ color: theme.colors.error }}>
-										!
-									</span>
-								) : (
-									<span
-										className="animate-pulse shrink-0 pt-0.5"
-										style={{ color: theme.colors.warning }}
-									>
-										●
-									</span>
-								)}
-								<span
-									className="break-words whitespace-pre-wrap opacity-80"
-									style={{ color: theme.colors.textMain }}
-								>
-									{log.text}
-								</span>
-							</div>
-						</div>
-					)}
-					{/* Special rendering for tool execution events (shown alongside thinking) */}
-					{log.source === 'tool' &&
-						(() => {
-							// Extract tool input details for display
-							const toolInput = log.metadata?.toolState?.input;
-							const toolSummary =
-								toolInput !== undefined && toolInput !== null
-									? summarizeToolInput(toolInput)
-									: null;
-							// Show the tool result once it has finished. Without this the
-							// compact tool log drops the output entirely (e.g. MCP calls
-							// like squash_repos that take no args render as a bare name).
-							const toolStatus = log.metadata?.toolState?.status;
-							const outputSummary =
-								toolStatus === 'completed' || toolStatus === 'failed' || toolStatus === 'error'
-									? summarizeToolOutput(log.metadata?.toolState?.output)
-									: null;
-
-							return (
-								<div
-									className="px-4 py-1.5 text-xs font-mono border-l-2"
-									style={{
-										color: theme.colors.textMain,
-										borderColor: theme.colors.accent,
-									}}
-								>
-									<div className="flex items-start gap-2">
-										<span
-											className="px-1.5 py-0.5 rounded shrink-0"
-											style={{
-												backgroundColor: `${theme.colors.accent}30`,
-												color: theme.colors.accent,
-											}}
-										>
-											{log.text}
-										</span>
-										{log.metadata?.toolState?.status === 'running' && (
-											<span
-												className="animate-pulse shrink-0 pt-0.5"
-												style={{ color: theme.colors.warning }}
-											>
-												●
-											</span>
-										)}
-										{log.metadata?.toolState?.status === 'completed' && (
-											<span className="shrink-0 pt-0.5" style={{ color: theme.colors.success }}>
-												✓
-											</span>
-										)}
-										{log.metadata?.toolState?.status === 'failed' && (
-											<span className="shrink-0 pt-0.5" style={{ color: theme.colors.error }}>
-												!
-											</span>
-										)}
-										{toolSummary?.description && (
-											<span
-												className="opacity-50 break-words"
-												style={{ color: theme.colors.textMain }}
-											>
-												{toolSummary.description}
-											</span>
-										)}
-									</div>
-									{toolSummary?.detail && (
-										<div
-											className="mt-1 ml-1 pl-2 opacity-70 break-words whitespace-pre-wrap border-l"
-											style={{
-												color: theme.colors.textMain,
-												borderColor: `${theme.colors.accent}40`,
-											}}
-										>
-											{toolSummary.detail}
-										</div>
-									)}
-									{outputSummary && (
-										<div
-											className="mt-1 ml-1 pl-2 opacity-60 break-words whitespace-pre-wrap border-l"
-											style={{
-												color: theme.colors.textMain,
-												borderColor: `${theme.colors.success}40`,
-											}}
-										>
-											{outputSummary}
-										</div>
-									)}
-								</div>
-							);
-						})()}
-					{!isHiddenProgressEntry(log) &&
-						log.source !== 'error' &&
-						log.source !== 'thinking' &&
-						log.source !== 'tool' &&
-						(hasNoMatches ? (
-							<div
-								className="flex items-center justify-center py-8 text-sm"
-								style={{ color: theme.colors.textDim }}
-							>
-								<span>No matches found for filter</span>
-							</div>
-						) : shouldCollapse && !isExpanded ? (
-							<div>
-								<div
-									className={`${isTerminal && log.source !== 'user' ? 'whitespace-pre text-sm' : 'whitespace-pre-wrap text-sm break-words'}`}
-									style={{
-										maxHeight: `${maxOutputLines * 1.5}em`,
-										overflow: isTerminal && log.source !== 'user' ? 'hidden' : 'hidden',
-										color: theme.colors.textMain,
-										fontFamily,
-										overflowWrap: isTerminal && log.source !== 'user' ? undefined : 'break-word',
-									}}
-								>
-									{isTerminal && log.source !== 'user' ? (
-										// Content sanitized with DOMPurify above
-										// Horizontal scroll for terminal output to preserve column alignment
-										<div
-											className="overflow-x-auto scrollbar-thin"
-											dangerouslySetInnerHTML={{ __html: displayHtmlContent }}
-										/>
-									) : isAIMode && !markdownEditMode ? (
-										// Collapsed markdown preview with rendered markdown
-										<MarkdownRenderer
-											content={displayText}
-											theme={theme}
-											onCopy={copyToClipboard}
-											enableBionifyReadingMode={bionifyReadingMode}
-											bionifyIntensity={bionifyIntensity}
-											bionifyAlgorithm={bionifyAlgorithm}
-											fileTree={fileTree}
-											cwd={cwd}
-											projectRoot={projectRoot}
-											onFileClick={onFileClick}
-											sshRemoteId={sshRemoteId}
-											chatLineBreaks
-											chatMath
-										/>
-									) : (
-										displayText
-									)}
-								</div>
-								<button
-									onClick={handleExpandToggle}
-									className="flex items-center gap-2 mt-2 text-xs px-3 py-1.5 rounded border hover:opacity-70 transition-opacity"
-									style={{
-										borderColor: theme.colors.border,
-										backgroundColor: theme.colors.bgActivity,
-										color: theme.colors.accent,
-									}}
-								>
-									<ChevronDown className="w-3 h-3" />
-									Show all {lineCount} lines
-								</button>
-							</div>
-						) : shouldCollapse && isExpanded ? (
-							<div>
-								<div
-									className={`${isTerminal && log.source !== 'user' ? 'whitespace-pre text-sm scrollbar-thin' : 'whitespace-pre-wrap text-sm break-words'}`}
-									style={{
-										maxHeight: '600px',
-										overflow: 'auto',
-										overscrollBehavior: 'contain',
-										color: theme.colors.textMain,
-										fontFamily,
-										overflowWrap: isTerminal && log.source !== 'user' ? undefined : 'break-word',
-									}}
-									onWheel={(e) => {
-										// Prevent scroll from propagating to parent when this container can scroll
-										const el = e.currentTarget;
-										const { scrollTop, scrollHeight, clientHeight } = el;
-										const atTop = scrollTop <= 0;
-										const atBottom = scrollTop + clientHeight >= scrollHeight - 1;
-
-										// Only stop propagation if we're not at the boundary we're scrolling towards
-										if ((e.deltaY < 0 && !atTop) || (e.deltaY > 0 && !atBottom)) {
-											e.stopPropagation();
-										}
-									}}
-								>
-									{isTerminal && log.source !== 'user' ? (
-										// Content sanitized with DOMPurify above
-										// Horizontal scroll for terminal output to preserve column alignment
-										<div dangerouslySetInnerHTML={{ __html: displayHtmlContent }} />
-									) : log.source === 'user' && isTerminal ? (
-										<div style={{ fontFamily }}>
-											<span style={{ color: theme.colors.accent }}>$ </span>
-											{filteredText}
-										</div>
-									) : log.aiCommand ? (
-										<div className="space-y-3">
-											<div
-												className="flex items-center gap-2 px-3 py-2 rounded-lg border"
-												style={{
-													backgroundColor: theme.colors.accent + '15',
-													borderColor: theme.colors.accent + '30',
-												}}
-											>
-												<span
-													className="font-mono font-bold text-sm"
-													style={{ color: theme.colors.accent }}
-												>
-													{log.aiCommand.command}:
-												</span>
-												<span className="text-sm" style={{ color: theme.colors.textMain }}>
-													{log.aiCommand.description}
-												</span>
-											</div>
-											<div>{linkifyNode(filteredText, theme)}</div>
-										</div>
-									) : isAIMode && !markdownEditMode ? (
-										// Expanded markdown rendering
-										<MarkdownRenderer
-											content={filteredText}
-											theme={theme}
-											onCopy={copyToClipboard}
-											enableBionifyReadingMode={bionifyReadingMode}
-											bionifyIntensity={bionifyIntensity}
-											bionifyAlgorithm={bionifyAlgorithm}
-											fileTree={fileTree}
-											cwd={cwd}
-											projectRoot={projectRoot}
-											onFileClick={onFileClick}
-											sshRemoteId={sshRemoteId}
-											chatLineBreaks
-											chatMath
-										/>
-									) : (
-										<div>{filteredText}</div>
-									)}
-								</div>
-								<button
-									onClick={handleExpandToggle}
-									className="flex items-center gap-2 mt-2 text-xs px-3 py-1.5 rounded border hover:opacity-70 transition-opacity"
-									style={{
-										borderColor: theme.colors.border,
-										backgroundColor: theme.colors.bgActivity,
-										color: theme.colors.accent,
-									}}
-								>
-									<ChevronUp className="w-3 h-3" />
-									Show less
-								</button>
-							</div>
-						) : (
-							<>
-								{isTerminal && log.source !== 'user' ? (
-									// Content sanitized with DOMPurify above
-									<div
-										className="whitespace-pre text-sm overflow-x-auto scrollbar-thin"
-										style={{
-											color: theme.colors.textMain,
-											fontFamily,
-											overscrollBehavior: 'contain',
-										}}
-										dangerouslySetInnerHTML={{ __html: displayHtmlContent }}
-									/>
-								) : log.source === 'user' && isTerminal ? (
-									<div
-										className="whitespace-pre-wrap text-sm break-words"
-										style={{ color: theme.colors.textMain, fontFamily }}
-									>
-										<span style={{ color: theme.colors.accent }}>$ </span>
-										{filteredText}
-									</div>
-								) : log.aiCommand ? (
-									<div className="space-y-3">
-										<div
-											className="flex items-center gap-2 px-3 py-2 rounded-lg border"
-											style={{
-												backgroundColor: theme.colors.accent + '15',
-												borderColor: theme.colors.accent + '30',
-											}}
-										>
-											<span
-												className="font-mono font-bold text-sm"
-												style={{ color: theme.colors.accent }}
-											>
-												{log.aiCommand.command}:
-											</span>
-											<span className="text-sm" style={{ color: theme.colors.textMain }}>
-												{log.aiCommand.description}
-											</span>
-										</div>
-										<div
-											className="whitespace-pre-wrap text-sm break-words"
-											style={{ color: theme.colors.textMain }}
-										>
-											{linkifyNode(filteredText, theme)}
-										</div>
-									</div>
-								) : isAIMode && !markdownEditMode ? (
-									// Rendered markdown for AI responses
-									<MarkdownRenderer
-										content={filteredText}
-										theme={theme}
-										onCopy={copyToClipboard}
-										enableBionifyReadingMode={bionifyReadingMode}
-										bionifyIntensity={bionifyIntensity}
-										bionifyAlgorithm={bionifyAlgorithm}
-										fileTree={fileTree}
-										cwd={cwd}
-										projectRoot={projectRoot}
-										onFileClick={onFileClick}
-										sshRemoteId={sshRemoteId}
-										chatLineBreaks
-										chatMath
-									/>
-								) : (
-									// Raw markdown source mode (show original text with markdown syntax visible)
-									<div
-										className="whitespace-pre-wrap text-sm break-words"
-										style={{ color: theme.colors.textMain }}
-									>
-										{filteredText}
-									</div>
-								)}
-							</>
-						))}
-					{/* Session-not-found recovery card. Rendered inline directly
-					    under the system message that announced the dead session. The
-					    card reads the tab from the store so we don't have to pass the
-					    full Session through LogItem (would defeat memoization). */}
-					{log.recoveryAction && (
-						<SessionRecoveryCardConnector
-							theme={theme}
-							sessionId={sessionId}
-							recoveryAction={log.recoveryAction}
-							isRecovering={!!isRecoveringSession}
-							recoveryError={sessionRecoveryError ?? null}
-							onRecover={(opts) => onSessionRecover?.(opts)}
-						/>
-					)}
-					{/* Cross-agent attribution now lives in the header at the TOP of the
-					    bubble (CrossAgentResponseHeader); no bottom pill is rendered here. */}
-					{/* Mode pill — shows which CLI captured this Claude turn (TUI Wrapper =
-					    maestro-p, claude -p = claude --print). "Dynamic " prefix indicates the
-					    session has Dynamic Mode enabled (auto-switching between the two).
-					    Suppressed on cross-agent entries so the attribution pill above
-					    replaces it. */}
-					{!crossAgent &&
-						isClaudeCode &&
-						log.source !== 'user' &&
-						(() => {
-							const { label, title } = getTokenSourcePill({
-								mode: log.renderStyle === 'text-stream' ? 'interactive' : 'api',
-								adaptive: isAdaptiveMode,
-							});
-							return (
-								<span
-									className="absolute bottom-2 left-1/2 -translate-x-1/2 text-[10px] px-1.5 py-0.5 rounded pointer-events-none select-none"
-									style={{
-										backgroundColor: `${theme.colors.accent}20`,
-										color: theme.colors.accent,
-										opacity: 0.7,
-									}}
-									title={title}
-								>
-									{label}
-								</span>
-							);
-						})()}
-					{/* Jump to top of this message - bottom left corner */}
-					<JumpToMessageTopButton
-						scrollContainerRef={scrollContainerRef}
-						messageRef={logItemRef}
-						theme={theme}
-					/>
-					{/* Action buttons - bottom right corner */}
-					<div
-						className="absolute bottom-2 right-2 flex items-center gap-1"
-						style={{ transition: 'opacity 0.15s ease-in-out' }}
-					>
-						{/* Markdown toggle button — available on both user and assistant
-						    messages in AI mode for consistent UX (#622). */}
-						{isAIMode && (
-							<button
-								onClick={onToggleMarkdownEditMode}
-								className="p-1.5 rounded opacity-0 group-hover:opacity-50 hover:!opacity-100"
-								style={{ color: markdownEditMode ? theme.colors.accent : theme.colors.textDim }}
-								title={
-									markdownEditMode
-										? `Show formatted (${formatShortcutKeys(['Meta', 'e'])})`
-										: `Show plain text (${formatShortcutKeys(['Meta', 'e'])})`
-								}
-							>
-								{markdownEditMode ? <Eye className="w-4 h-4" /> : <FileText className="w-4 h-4" />}
-							</button>
-						)}
-						{/* Replay button for user messages in AI mode */}
-						{isUserMessage && isAIMode && onReplayMessage && (
-							<button
-								onClick={() => onReplayMessage(log.text, log.images)}
-								className="p-1.5 rounded opacity-0 group-hover:opacity-50 hover:!opacity-100"
-								style={{ color: theme.colors.textDim }}
-								title="Replay message"
-							>
-								<RotateCcw className="w-3.5 h-3.5" />
-							</button>
-						)}
-						{/* Copy to Clipboard Button */}
-						<button
-							onClick={() => copyToClipboard(log.text)}
-							className="p-1.5 rounded opacity-0 group-hover:opacity-50 hover:!opacity-100"
-							style={{ color: theme.colors.textDim }}
-							title="Copy to clipboard"
-						>
-							<Copy className="w-3.5 h-3.5" />
-						</button>
-						{/* Save to File Button - only for AI responses */}
-						{log.source !== 'user' && isAIMode && onSaveToFile && (
-							<button
-								onClick={() => onSaveToFile(log.text)}
-								className="p-1.5 rounded opacity-0 group-hover:opacity-50 hover:!opacity-100"
-								style={{ color: theme.colors.textDim }}
-								title="Save to file"
-							>
-								<Save className="w-3.5 h-3.5" />
-							</button>
-						)}
-						{/* Fork conversation — user messages and AI responses (source='stdout' in AI mode, or 'ai' if ever set) */}
-						{(log.source === 'user' || log.source === 'ai' || log.source === 'stdout') &&
-							isAIMode &&
-							onForkConversation && (
-								<button
-									onClick={() => onForkConversation(log.id)}
-									className="p-1.5 rounded opacity-0 group-hover:opacity-50 hover:!opacity-100"
-									style={{ color: theme.colors.textDim }}
-									title="Fork conversation from here"
-								>
-									<GitFork className="w-3.5 h-3.5" />
-								</button>
-							)}
-						{/* Publish to GitHub Gist - only for AI responses when gh CLI available */}
-						{log.source !== 'user' && isAIMode && ghCliAvailable && onPublishGist && (
-							<button
-								onClick={() => onPublishGist(log.text, log.id)}
-								className={`p-1.5 rounded hover:!opacity-100 ${
-									publishedGistUrl ? 'opacity-100' : 'opacity-0 group-hover:opacity-50'
-								}`}
-								style={{
-									color: publishedGistUrl ? theme.colors.accent : theme.colors.textDim,
-								}}
-								title={
-									publishedGistUrl
-										? `Published as Gist: ${publishedGistUrl}`
-										: 'Publish as GitHub Gist'
-								}
-							>
-								<Share2 className="w-3.5 h-3.5" />
-							</button>
-						)}
-						{/* Delete button for user messages (both AI and terminal modes) */}
-						{log.source === 'user' &&
-							onDeleteLog &&
-							(deleteConfirmLogId === log.id ? (
-								<div
-									className="flex items-center gap-1 p-1 rounded border"
-									style={{
-										backgroundColor: theme.colors.bgSidebar,
-										borderColor: theme.colors.error,
-									}}
-								>
-									<span className="text-xs px-1" style={{ color: theme.colors.error }}>
-										Delete?
-									</span>
-									<button
-										onClick={() => {
-											const nextIndex = onDeleteLog(log.id);
-											onSetDeleteConfirmLogId(null);
-											if (nextIndex !== null && nextIndex >= 0) {
-												setTimeout(() => {
-													const container = scrollContainerRef.current;
-													const items = container?.querySelectorAll('[data-log-index]');
-													const targetItem = items?.[nextIndex] as HTMLElement;
-													if (targetItem && container) {
-														container.scrollTop = targetItem.offsetTop;
-													}
-												}, 50);
-											}
-										}}
-										className="px-2 py-0.5 rounded text-xs font-medium hover:opacity-80"
-										style={{ backgroundColor: theme.colors.error, color: '#fff' }}
-									>
-										Yes
-									</button>
-									<button
-										onClick={() => onSetDeleteConfirmLogId(null)}
-										className="px-2 py-0.5 rounded text-xs hover:opacity-80"
-										style={{ color: theme.colors.textDim }}
-									>
-										No
-									</button>
-								</div>
-							) : (
-								<button
-									onClick={() => onSetDeleteConfirmLogId(log.id)}
-									className="p-1.5 rounded opacity-0 group-hover:opacity-50 hover:!opacity-100 transition-opacity"
-									style={{ color: theme.colors.textDim }}
-									title={isAIMode ? 'Delete message and response' : 'Delete command and output'}
-								>
-									<Trash2 className="w-3.5 h-3.5" />
-								</button>
-							))}
-						{/* Read-only mode indicator for messages sent in read-only/plan mode */}
-						{isUserMessage && isAIMode && log.readOnly && (
-							<span title="Sent in read-only mode" className="flex items-center">
-								<Eye
-									className="w-3.5 h-3.5"
-									style={{ color: theme.colors.warning, opacity: 0.7 }}
-								/>
-							</span>
-						)}
-						{/* Force parallel indicator for messages sent via Cmd+Shift+Enter */}
-						{isUserMessage && isAIMode && log.forceParallel && (
-							<span
-								title="Sent via forced parallel execution (bypassed queue)"
-								className="flex items-center"
-							>
-								<Hammer
-									className="w-3.5 h-3.5"
-									style={{ color: theme.colors.warning, opacity: 0.7 }}
-								/>
-							</span>
-						)}
-						{/* Delivery checkmark for user messages in AI mode - positioned at the end */}
-						{isUserMessage && isAIMode && log.delivered && (
-							<span title="Message delivered" className="flex items-center">
-								<Check
-									className="w-3.5 h-3.5"
-									style={{ color: theme.colors.success, opacity: 0.6 }}
-								/>
-							</span>
-						)}
-					</div>
-				</div>
-			</div>
-		);
-	},
-	(prevProps, nextProps) => {
-		// Custom comparison - only re-render if these specific props change
-		// IMPORTANT: Include ALL props that affect visual rendering
-		return (
-			prevProps.log.id === nextProps.log.id &&
-			prevProps.log.text === nextProps.log.text &&
-			prevProps.log.delivered === nextProps.log.delivered &&
-			prevProps.log.readOnly === nextProps.log.readOnly &&
-			prevProps.log.forceParallel === nextProps.log.forceParallel &&
-			prevProps.log.renderStyle === nextProps.log.renderStyle &&
-			prevProps.log.metadata?.hiddenProgress === nextProps.log.metadata?.hiddenProgress &&
-			prevProps.log.metadata?.toolState?.status === nextProps.log.metadata?.toolState?.status &&
-			// Cross-agent streaming flips true->false on the final chunk (which may
-			// carry no new text), so the pill/glow needs this to settle.
-			prevProps.log.metadata?.crossAgent?.streaming ===
-				nextProps.log.metadata?.crossAgent?.streaming &&
-			// A terminal error chunk may add `error` without changing text; the
-			// red-tinted bubble variant depends on it.
-			prevProps.log.metadata?.crossAgent?.error === nextProps.log.metadata?.crossAgent?.error &&
-			prevProps.isExpanded === nextProps.isExpanded &&
-			prevProps.localFilterQuery === nextProps.localFilterQuery &&
-			prevProps.filterMode.mode === nextProps.filterMode.mode &&
-			prevProps.filterMode.regex === nextProps.filterMode.regex &&
-			prevProps.activeLocalFilter === nextProps.activeLocalFilter &&
-			prevProps.deleteConfirmLogId === nextProps.deleteConfirmLogId &&
-			prevProps.theme === nextProps.theme &&
-			prevProps.maxOutputLines === nextProps.maxOutputLines &&
-			prevProps.markdownEditMode === nextProps.markdownEditMode &&
-			prevProps.bionifyReadingMode === nextProps.bionifyReadingMode &&
-			prevProps.bionifyIntensity === nextProps.bionifyIntensity &&
-			prevProps.bionifyAlgorithm === nextProps.bionifyAlgorithm &&
-			prevProps.fontFamily === nextProps.fontFamily &&
-			prevProps.userMessageAlignment === nextProps.userMessageAlignment &&
-			prevProps.ghCliAvailable === nextProps.ghCliAvailable &&
-			prevProps.onForkConversation === nextProps.onForkConversation &&
-			prevProps.publishedGistUrl === nextProps.publishedGistUrl
-		);
-	}
-);
-
-LogItemComponent.displayName = 'LogItemComponent';
-
-interface TerminalOutputProps {
-	session: Session;
-	theme: Theme;
-	fontFamily: string;
-	activeFocus: FocusArea;
-	outputSearchOpen: boolean;
-	outputSearchQuery: string;
-	outputSearchRegex: boolean;
-	setOutputSearchOpen: (open: boolean) => void;
-	setOutputSearchQuery: (query: string) => void;
-	setOutputSearchRegex: (regex: boolean) => void;
-	setActiveFocus: (focus: FocusArea) => void;
-	setLightboxImage: (
-		image: string | null,
-		contextImages?: string[],
-		source?: 'staged' | 'history'
-	) => void;
-	inputRef: React.RefObject<HTMLTextAreaElement>;
-	logsEndRef: React.RefObject<HTMLDivElement>;
-	maxOutputLines: number;
-	onDeleteLog?: (logId: string) => number | null; // Returns the index to scroll to after deletion
-	onRemoveQueuedItem?: (itemId: string) => void; // Callback to remove a queued item from execution queue
-	onTogglePauseQueuedItem?: (itemId: string) => void; // Callback to toggle held/paused state of a queued item
-	onEditQueuedItem?: (itemId: string, patch: { text: string; images: string[] }) => void; // Edit a queued message's text + images
-	onReorderQueuedItem?: (fromIndex: number, toIndex: number, tabId?: string) => void; // Reorder a queued item within the active tab's queue
-	onForceSendQueuedItem?: (itemId: string) => void; // Callback to Force Send a queued item (parallel execution)
-	forcedParallelEnabled?: boolean; // Whether forcedParallelExecution setting is on (gates Force Send button)
-	getForceSendContext?: (
-		item: QueuedItem
-	) => { targetTabBusy: boolean; otherBusyTabs: { id: string; displayName: string }[] } | null;
-	onInterrupt?: () => void; // Callback to interrupt the current process
-	onScrollPositionChange?: (scrollTop: number) => void; // Callback to save scroll position
-	onAtBottomChange?: (isAtBottom: boolean) => void; // Callback when user scrolls to/away from bottom
-	initialScrollTop?: number; // Initial scroll position to restore
-	markdownEditMode: boolean; // Whether to show raw markdown or rendered markdown for AI responses
-	setMarkdownEditMode: (value: boolean) => void; // Toggle markdown mode
-	onReplayMessage?: (text: string, images?: string[]) => void; // Replay a user message
-	onForkConversation?: (logId: string) => void; // Fork conversation from a specific message
-	fileTree?: FileNode[]; // File tree for linking file references
-	cwd?: string; // Current working directory for proximity-based matching
-	projectRoot?: string; // Project root absolute path for converting absolute paths to relative
-	onFileClick?: (path: string) => void; // Callback when a file link is clicked
-	onShowErrorDetails?: (error: AgentError) => void; // Callback to show the error modal (for error log entries)
-	onFileSaved?: () => void; // Callback when markdown content is saved to file (e.g., to refresh file list)
-	userMessageAlignment?: 'left' | 'right'; // User message bubble alignment (default: right)
-	ghCliAvailable?: boolean; // Whether gh CLI is available for gist publishing
-	onPublishMessageGist?: (text: string, messageId?: string) => void; // Callback to publish a single message as a gist
-	onOpenInTab?: (file: {
-		path: string;
-		name: string;
-		content: string;
-		sshRemoteId?: string;
-	}) => void; // Callback to open saved file in a tab
-	// In-place recovery from session_not_found errors. Invoked by the
-	// SessionRecoveryCard that renders inside system log entries carrying a
-	// `recoveryAction` payload.
-	onSessionRecover?: (opts: {
-		sessionId: string;
-		tabId: string;
-		lastUserPrompt: string;
-		groomContext: boolean;
-	}) => void;
-	isRecoveringSession?: boolean;
-	sessionRecoveryError?: string | null;
-}
-
-// PERFORMANCE: Wrap in React.memo to prevent re-renders when parent re-renders
-// but TerminalOutput's props haven't changed. This is critical because TerminalOutput
-// can render many log entries and is expensive to re-render.
-export const TerminalOutput = memo(
-	forwardRef<HTMLDivElement, TerminalOutputProps>((props, ref) => {
-		const {
-			session,
-			theme,
-			fontFamily,
-			activeFocus: _activeFocus,
-			outputSearchOpen,
-			outputSearchQuery,
-			outputSearchRegex,
-			setOutputSearchOpen,
-			setOutputSearchQuery,
-			setOutputSearchRegex,
-			setActiveFocus,
-			setLightboxImage,
-			inputRef,
-			logsEndRef,
-			maxOutputLines,
-			onDeleteLog,
-			onRemoveQueuedItem,
-			onTogglePauseQueuedItem,
-			onEditQueuedItem,
-			onReorderQueuedItem,
-			onForceSendQueuedItem,
-			forcedParallelEnabled,
-			getForceSendContext,
-			onInterrupt: _onInterrupt,
-			onScrollPositionChange,
-			onAtBottomChange,
-			initialScrollTop,
-			markdownEditMode,
-			setMarkdownEditMode,
-			onReplayMessage,
-			onForkConversation,
-			fileTree,
-			cwd,
-			projectRoot,
-			onFileClick,
-			onShowErrorDetails,
-			onFileSaved,
-			userMessageAlignment = 'right',
-			onOpenInTab,
-			ghCliAvailable,
-			onPublishMessageGist,
-			onSessionRecover,
-			isRecoveringSession,
-			sessionRecoveryError,
-		} = props;
-		const globalBionifyReadingMode = useSettingsStore((s) => s.bionifyReadingMode);
-		const globalBionifyIntensity = useSettingsStore((s) => s.bionifyIntensity);
-		const publishedGists = useMessageGistStore((s) => s.published);
-		const globalBionifyAlgorithm = useSettingsStore((s) => s.bionifyAlgorithm);
-
-		// Use the forwarded ref if provided, otherwise create a local one
-		const localRef = useRef<HTMLDivElement>(null);
-		const terminalOutputRef = (ref as React.RefObject<HTMLDivElement>) || localRef;
-
-		// Scroll container ref for native scrolling
-		const scrollContainerRef = useRef<HTMLDivElement>(null);
-
-		// Track which log entries are expanded (by log ID)
-		const [expandedLogs, setExpandedLogs] = useState<Set<string>>(new Set());
-		// Use a ref to access current value without recreating LogItem callback
-		const expandedLogsRef = useRef(expandedLogs);
-		expandedLogsRef.current = expandedLogs;
-		// Counter to force re-render of LogItem when expanded state changes
-		const [_expandedTrigger, setExpandedTrigger] = useState(0);
-
-		// Track local filters per log entry (log ID -> filter query)
-		const [localFilters, setLocalFilters] = useState<Map<string, string>>(new Map());
-		// Use refs to access current values without recreating LogItem callback
-		const localFiltersRef = useRef(localFilters);
-		localFiltersRef.current = localFilters;
-		const [activeLocalFilter, setActiveLocalFilter] = useState<string | null>(null);
-		const activeLocalFilterRef = useRef(activeLocalFilter);
-		activeLocalFilterRef.current = activeLocalFilter;
-		// Counter to force re-render when local filter state changes
-		const [_filterTrigger, setFilterTrigger] = useState(0);
-
-		// Track filter modes per log entry (log ID -> {mode: 'include'|'exclude', regex: boolean})
-		const [filterModes, setFilterModes] = useState<
-			Map<string, { mode: 'include' | 'exclude'; regex: boolean }>
-		>(new Map());
-		const filterModesRef = useRef(filterModes);
-		filterModesRef.current = filterModes;
-
-		// Delete confirmation state
-		const [deleteConfirmLogId, setDeleteConfirmLogId] = useState<string | null>(null);
-		const deleteConfirmLogIdRef = useRef(deleteConfirmLogId);
-		deleteConfirmLogIdRef.current = deleteConfirmLogId;
-		// Counter to force re-render when delete confirmation changes
-		const [_deleteConfirmTrigger, _setDeleteConfirmTrigger] = useState(0);
-
-		// Save markdown modal state
-		const [saveModalContent, setSaveModalContent] = useState<string | null>(null);
-
-		// New message indicator state
-		const [isAtBottom, setIsAtBottom] = useState(true);
-		const [hasNewMessages, setHasNewMessages] = useState(false);
-		const [newMessageCount, setNewMessageCount] = useState(0);
-		const lastLogCountRef = useRef(0);
-		// Track previous isAtBottom to detect changes for callback
-		const prevIsAtBottomRef = useRef(true);
-		// Ref mirror of isAtBottom for MutationObserver closure (avoids stale state)
-		const isAtBottomRef = useRef(true);
-		isAtBottomRef.current = isAtBottom;
-		// Track whether auto-scroll is paused because user scrolled up (state so button re-renders)
-		const [autoScrollPaused, setAutoScrollPaused] = useState(false);
-		// Ref mirror of autoScrollPaused for the MutationObserver closure so a freshly
-		// restored scroll position can suppress auto-scroll synchronously, before the
-		// state-driven re-render re-runs the observer effect (avoids a one-frame yank).
-		const autoScrollPausedRef = useRef(false);
-		autoScrollPausedRef.current = autoScrollPaused;
-		// Guard flag: prevents the scroll handler from pausing auto-scroll
-		// during programmatic scrollTo() calls from the MutationObserver effect.
-		const isProgrammaticScrollRef = useRef(false);
-
-		// Track read state per tab - stores the log count when user scrolled to bottom
-		const tabReadStateRef = useRef<Map<string, number>>(new Map());
-
-		// Throttle timer ref for scroll position saves
-		const scrollSaveTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-		// Track if initial scroll restore has been done
-		const hasRestoredScrollRef = useRef(false);
-
-		// Get active tab ID for resetting state on tab switch
-		const activeTabId = session.activeTabId;
-
-		// Copy text to clipboard with center flash
-		const copyToClipboard = useCallback(async (text: string) => {
-			const ok = await safeClipboardWrite(text);
-			if (ok) {
-				flashCopiedToClipboard(text);
-			}
-		}, []);
-
-		// Open save modal for markdown content
-		const handleSaveToFile = useCallback((text: string) => {
-			setSaveModalContent(text);
-		}, []);
-
-		// Layer stack integration for search overlay
-		const { registerLayer, unregisterLayer, updateLayerHandler } = useLayerStack();
-		const layerIdRef = useRef<string>();
-
-		// Register layer when search is open
-		useEffect(() => {
-			if (outputSearchOpen) {
-				layerIdRef.current = registerLayer({
-					type: 'overlay',
-					priority: MODAL_PRIORITIES.SLASH_AUTOCOMPLETE, // Use same priority as slash autocomplete (low priority)
-					blocksLowerLayers: false,
-					capturesFocus: true,
-					focusTrap: 'none',
-					onEscape: () => {
-						setOutputSearchOpen(false);
-						setOutputSearchQuery('');
-						terminalOutputRef.current?.focus();
-					},
-					allowClickOutside: true,
-					ariaLabel: 'Output Search',
-				});
-
-				return () => {
-					if (layerIdRef.current) {
-						unregisterLayer(layerIdRef.current);
-					}
-				};
-			}
-		}, [outputSearchOpen, registerLayer, unregisterLayer]);
-
-		// Update the handler when dependencies change
-		useEffect(() => {
-			if (outputSearchOpen && layerIdRef.current) {
-				updateLayerHandler(layerIdRef.current, () => {
-					setOutputSearchOpen(false);
-					setOutputSearchQuery('');
-					terminalOutputRef.current?.focus();
-				});
-			}
-		}, [outputSearchOpen, updateLayerHandler]);
-
-		// Search match navigation state (populated by effect below after filteredLogs is defined)
-		const [currentMatchIndex, setCurrentMatchIndex] = useState(0);
-		const [totalMatches, setTotalMatches] = useState(0);
-		const [regexError, setRegexError] = useState<string | null>(null);
-		const matchRangesRef = useRef<Range[]>([]);
-
-		const toggleExpanded = useCallback((logId: string) => {
-			setExpandedLogs((prev) => {
-				const newSet = new Set(prev);
-				if (newSet.has(logId)) {
-					newSet.delete(logId);
-				} else {
-					newSet.add(logId);
-				}
-				return newSet;
-			});
-			// Trigger re-render after state update
-			setExpandedTrigger((t) => t + 1);
-		}, []);
-
-		const toggleLocalFilter = useCallback((logId: string) => {
-			setActiveLocalFilter((prev) => (prev === logId ? null : logId));
-			setFilterTrigger((t) => t + 1);
-		}, []);
-
-		const setLocalFilterQuery = useCallback((logId: string, query: string) => {
-			setLocalFilters((prev) => {
-				const newMap = new Map(prev);
-				if (query) {
-					newMap.set(logId, query);
-				} else {
-					newMap.delete(logId);
-				}
-				return newMap;
-			});
-		}, []);
-
-		// Callback to update filter mode for a log entry
-		const setFilterModeForLog = useCallback(
-			(
-				logId: string,
-				update: (current: { mode: 'include' | 'exclude'; regex: boolean }) => {
-					mode: 'include' | 'exclude';
-					regex: boolean;
-				}
-			) => {
-				setFilterModes((prev) => {
-					const newMap = new Map(prev);
-					const current = newMap.get(logId) || { mode: 'include' as const, regex: false };
-					newMap.set(logId, update(current));
-					return newMap;
-				});
-			},
-			[]
-		);
-
-		// Callback to clear local filter for a log entry
-		const clearLocalFilter = useCallback(
-			(logId: string) => {
-				setActiveLocalFilter(null);
-				setLocalFilterQuery(logId, '');
-				setFilterModes((prev) => {
-					const newMap = new Map(prev);
-					newMap.delete(logId);
-					return newMap;
-				});
-			},
-			[setLocalFilterQuery]
-		);
-
-		// Callback to toggle markdown mode
-		const toggleMarkdownEditMode = useCallback(() => {
-			setMarkdownEditMode(!markdownEditMode);
-		}, [markdownEditMode, setMarkdownEditMode]);
-
-		// Auto-focus on search input when opened
-		useEffect(() => {
-			if (outputSearchOpen) {
-				terminalOutputRef.current?.querySelector('input')?.focus();
-			}
-		}, [outputSearchOpen]);
-
-		// Create ANSI converter with theme-aware colors
-		const ansiConverter = useMemo(() => {
-			const c = theme.colors;
-			return new Convert({
-				fg: c.textMain,
-				bg: c.bgMain,
-				newline: false,
-				escapeXML: true,
-				stream: false,
-				colors: {
-					0: c.ansiBlack ?? c.textMain,
-					1: c.ansiRed ?? c.error,
-					2: c.ansiGreen ?? c.success,
-					3: c.ansiYellow ?? c.warning,
-					4: c.ansiBlue ?? c.accent,
-					5: c.ansiMagenta ?? c.accentDim,
-					6: c.ansiCyan ?? c.accent,
-					7: c.ansiWhite ?? c.textDim,
-					8: c.ansiBrightBlack ?? c.textDim,
-					9: c.ansiBrightRed ?? c.error,
-					10: c.ansiBrightGreen ?? c.success,
-					11: c.ansiBrightYellow ?? c.warning,
-					12: c.ansiBrightBlue ?? c.accent,
-					13: c.ansiBrightMagenta ?? c.accentText,
-					14: c.ansiBrightCyan ?? c.accentText,
-					15: c.ansiBrightWhite ?? c.textMain,
-				},
-			});
-		}, [theme]);
-
-		// PERF: Memoize active tab lookup to avoid O(n) .find() on every render
-		const activeTab = useMemo(() => getActiveTab(session), [session.aiTabs, session.activeTabId]);
-
-		// PERF: Memoize activeLogs to provide stable reference for collapsedLogs dependency
-		// TerminalOutput only handles AI mode; terminal mode renders via TerminalView
-		const activeLogs = useMemo((): LogEntry[] => activeTab?.logs ?? [], [activeTab?.logs]);
-
-		// In AI mode, collapse consecutive local response entries into single blocks
-		// so each user message gets one bubble. Tool/thinking entries, Agent
-		// Resilience outage markers, and cross-agent (@mention) replies stay
-		// standalone (see collapseAiResponseLogs).
-		const collapsedLogs = useMemo(() => collapseAiResponseLogs(activeLogs), [activeLogs]);
-
-		// PERF: Debounce search query so the highlight pass doesn't run on every keystroke
-		const debouncedSearchQuery = useDebouncedValue(outputSearchQuery, 150);
-
-		// Search no longer filters logs — all logs stay visible. Matches are highlighted and
-		// navigated inline via CSS Custom Highlight API (see highlight effect below).
-		const filteredLogs = collapsedLogs;
-
-		// ============================================================================
-		// Search match navigation (CSS Custom Highlight API)
-		// ============================================================================
-		// Pattern mirrors `useFilePreviewSearch.ts` markdown path: walks text nodes in the
-		// scroll container, builds Range objects for matches, and uses the Custom Highlight
-		// API to paint them without mutating the DOM. The "current" match gets a separate
-		// highlight with accent color; prev/next navigation just moves the index.
-		useEffect(() => {
-			const query = debouncedSearchQuery.trim();
-			const clearHighlights = () => {
-				if ('highlights' in CSS) {
-					(CSS as unknown as { highlights: Map<string, unknown> }).highlights.delete(
-						'terminal-search-all'
-					);
-					(CSS as unknown as { highlights: Map<string, unknown> }).highlights.delete(
-						'terminal-search-current'
-					);
-				}
-			};
-			if (!outputSearchOpen || !query) {
-				clearHighlights();
-				matchRangesRef.current = [];
-				setTotalMatches(0);
-				setCurrentMatchIndex(0);
-				setRegexError(null);
-				return;
-			}
-
-			const container = scrollContainerRef.current;
-			if (!container) return;
-
-			// Build the match regex — plain text is escaped; regex mode is validated.
-			let regex: RegExp;
-			try {
-				if (outputSearchRegex) {
-					regex = new RegExp(query, 'gi');
-				} else {
-					const escaped = query.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-					regex = new RegExp(escaped, 'gi');
-				}
-				setRegexError(null);
-			} catch (err) {
-				setRegexError(err instanceof Error ? err.message : 'Invalid regex');
-				clearHighlights();
-				matchRangesRef.current = [];
-				setTotalMatches(0);
-				return;
-			}
-
-			// Walk text nodes, collect Range objects for each match.
-			const ranges: Range[] = [];
-			const walker = document.createTreeWalker(container, NodeFilter.SHOW_TEXT);
-			let textNode: Node | null = walker.nextNode();
-			while (textNode !== null) {
-				const text = textNode.textContent || '';
-				if (text) {
-					regex.lastIndex = 0;
-					let m: RegExpExecArray | null = regex.exec(text);
-					while (m !== null) {
-						if (m[0].length === 0) {
-							regex.lastIndex++;
-						} else {
-							const range = document.createRange();
-							range.setStart(textNode, m.index);
-							range.setEnd(textNode, m.index + m[0].length);
-							ranges.push(range);
-						}
-						m = regex.exec(text);
-					}
-				}
-				textNode = walker.nextNode();
-			}
-
-			matchRangesRef.current = ranges;
-			setTotalMatches(ranges.length);
-			setCurrentMatchIndex((prev) => (ranges.length === 0 ? 0 : Math.min(prev, ranges.length - 1)));
-
-			if (!('highlights' in CSS) || ranges.length === 0) {
-				clearHighlights();
-				return;
-			}
-			const Highlight = (window as unknown as { Highlight: new (...r: Range[]) => unknown })
-				.Highlight;
-			const highlights = (CSS as unknown as { highlights: Map<string, unknown> }).highlights;
-			highlights.set('terminal-search-all', new Highlight(...ranges));
-			// Current highlight is applied by the separate effect below so navigation doesn't
-			// require re-walking the DOM.
-
-			return clearHighlights;
-		}, [debouncedSearchQuery, outputSearchRegex, outputSearchOpen, filteredLogs]);
-
-		// Update the "current" highlight and scroll it into view when index changes.
-		useEffect(() => {
-			if (!('highlights' in CSS)) return;
-			const highlights = (CSS as unknown as { highlights: Map<string, unknown> }).highlights;
-			const ranges = matchRangesRef.current;
-			if (ranges.length === 0 || currentMatchIndex < 0 || currentMatchIndex >= ranges.length) {
-				highlights.delete('terminal-search-current');
-				return;
-			}
-			const current = ranges[currentMatchIndex];
-			const Highlight = (window as unknown as { Highlight: new (...r: Range[]) => unknown })
-				.Highlight;
-			highlights.set('terminal-search-current', new Highlight(current));
-
-			// Scroll the current match into view, centered in the scroll container.
-			const scrollParent = scrollContainerRef.current;
-			const rect = current.getBoundingClientRect();
-			if (scrollParent && rect.height > 0) {
-				const parentRect = scrollParent.getBoundingClientRect();
-				const offset = rect.top - parentRect.top + scrollParent.scrollTop;
-				const targetScroll = offset - scrollParent.clientHeight / 2 + rect.height / 2;
-				scrollParent.scrollTo({ top: Math.max(0, targetScroll), behavior: 'smooth' });
-			}
-		}, [currentMatchIndex, totalMatches]);
-
-		const goToNextMatch = useCallback(() => {
-			setCurrentMatchIndex((i) => {
-				if (totalMatches === 0) return 0;
-				return (i + 1) % totalMatches;
-			});
-		}, [totalMatches]);
-
-		const goToPrevMatch = useCallback(() => {
-			setCurrentMatchIndex((i) => {
-				if (totalMatches === 0) return 0;
-				return (i - 1 + totalMatches) % totalMatches;
-			});
-		}, [totalMatches]);
-
-		// PERF: Throttle scroll handler to reduce state updates (4ms = ~240fps for smooth scrollbar)
-		// The actual logic is in handleScrollInner, wrapped with useThrottledCallback
-		const handleScrollInner = useCallback(() => {
-			if (!scrollContainerRef.current) return;
-			const { scrollTop, scrollHeight, clientHeight } = scrollContainerRef.current;
-			// Consider "at bottom" if within 50px of the bottom
-			const atBottom = scrollHeight - scrollTop - clientHeight < 50;
-			setIsAtBottom(atBottom);
-
-			// Notify parent when isAtBottom changes (for hasUnread logic)
-			if (atBottom !== prevIsAtBottomRef.current) {
-				prevIsAtBottomRef.current = atBottom;
-				onAtBottomChange?.(atBottom);
-			}
-
-			// Clear new message indicator when user scrolls to bottom
-			if (atBottom) {
-				setHasNewMessages(false);
-				setNewMessageCount(0);
-				// Resume auto-scroll when user scrolls back to bottom
-				setAutoScrollPaused(false);
-				// Save read state for current tab
-				if (activeTabId) {
-					tabReadStateRef.current.set(activeTabId, filteredLogs.length);
-				}
-			} else {
-				if (isProgrammaticScrollRef.current) {
-					// This scroll event was triggered by our own scrollTo() call —
-					// consume the guard flag here inside the throttled handler to avoid
-					// the race where queueMicrotask clears the flag before a deferred
-					// throttled invocation fires (throttle delay is 16ms > microtask).
-					isProgrammaticScrollRef.current = false;
-				} else {
-					// Genuine user scroll away from bottom — pause auto-scroll
-					setAutoScrollPaused(true);
-				}
-			}
-
-			// Throttled scroll position save (200ms)
-			if (onScrollPositionChange) {
-				if (scrollSaveTimerRef.current) {
-					clearTimeout(scrollSaveTimerRef.current);
-				}
-				scrollSaveTimerRef.current = setTimeout(() => {
-					onScrollPositionChange(scrollTop);
-					scrollSaveTimerRef.current = null;
-				}, 200);
-			}
-		}, [activeTabId, filteredLogs.length, onScrollPositionChange, onAtBottomChange]);
-
-		// PERF: Throttle at 16ms (60fps) instead of 4ms to reduce state updates during scroll
-		const handleScroll = useThrottledCallback(handleScrollInner, 16);
-
-		// Restore read state when switching tabs
-		useEffect(() => {
-			if (!activeTabId) {
-				// Terminal mode - just reset
-				setHasNewMessages(false);
-				setNewMessageCount(0);
-				setIsAtBottom(true);
-				lastLogCountRef.current = filteredLogs.length;
-				return;
-			}
-
-			// Restore saved read state for this tab
-			const savedReadCount = tabReadStateRef.current.get(activeTabId);
-			const currentCount = filteredLogs.length;
-
-			if (savedReadCount !== undefined) {
-				// Tab was visited before - check for new messages since last read
-				const unreadCount = currentCount - savedReadCount;
-				if (unreadCount > 0) {
-					setHasNewMessages(true);
-					setNewMessageCount(unreadCount);
-					setIsAtBottom(false);
-				} else {
-					setHasNewMessages(false);
-					setNewMessageCount(0);
-					setIsAtBottom(true);
-				}
-			} else {
-				// First visit to this tab - mark all as read
-				tabReadStateRef.current.set(activeTabId, currentCount);
-				setHasNewMessages(false);
-				setNewMessageCount(0);
-				setIsAtBottom(true);
-			}
-
-			lastLogCountRef.current = currentCount;
-		}, [activeTabId]); // Only run when tab changes, not when filteredLogs changes
-
-		// Detect new messages when user is not at bottom (while staying on same tab).
-		// NOTE: This intentionally uses filteredLogs.length (not the MutationObserver) because
-		// unread badge counts should only increment on NEW log entries, not on in-place text
-		// updates (thinking stream growth). The MutationObserver handles scroll triggering;
-		// this effect handles the unread badge.
-		useEffect(() => {
-			const currentCount = filteredLogs.length;
-			if (currentCount > lastLogCountRef.current) {
-				// Check actual scroll position, not just state (state may be stale)
-				const container = scrollContainerRef.current;
-				let actuallyAtBottom = isAtBottom;
-				if (container) {
-					const { scrollTop, scrollHeight, clientHeight } = container;
-					actuallyAtBottom = scrollHeight - scrollTop - clientHeight < 50;
-				}
-
-				if (!actuallyAtBottom) {
-					const newCount = currentCount - lastLogCountRef.current;
-					setHasNewMessages(true);
-					setNewMessageCount((prev) => prev + newCount);
-					// Update isAtBottom state to match reality
-					setIsAtBottom(false);
-				} else {
-					// At bottom, update read state
-					if (activeTabId) {
-						tabReadStateRef.current.set(activeTabId, currentCount);
-					}
-				}
-			}
-			lastLogCountRef.current = currentCount;
-		}, [filteredLogs.length, isAtBottom, activeTabId]);
-
-		// Auto-scroll to bottom when DOM content changes in the scroll container.
-		// Uses MutationObserver to detect ALL content mutations — new nodes (log entries),
-		// text changes (thinking stream growth), and attribute changes (tool status updates).
-		// This replaces the previous filteredLogs.length dependency, which missed in-place
-		// text updates during thinking/tool streaming (GitHub issue #402).
-		useEffect(() => {
-			const container = scrollContainerRef.current;
-			if (!container) return;
-
-			const shouldAutoScroll = () => !autoScrollPausedRef.current || isAtBottomRef.current;
-
-			const scrollToBottom = () => {
-				if (!scrollContainerRef.current) return;
-				requestAnimationFrame(() => {
-					if (scrollContainerRef.current) {
-						// Set guard flag BEFORE scrollTo — the throttled scroll handler
-						// checks this flag and consumes it (clears it) when it fires,
-						// preventing the programmatic scroll from being misinterpreted
-						// as a user scroll-up that should pause auto-scroll.
-						isProgrammaticScrollRef.current = true;
-						scrollContainerRef.current.scrollTo({
-							top: scrollContainerRef.current.scrollHeight,
-							behavior: 'auto',
-						});
-						// Fallback: if scrollTo is a no-op (already at bottom), the browser
-						// won't fire a scroll event, so the handler never consumes the guard.
-						// Clear it after 32ms (2x the 16ms throttle window) to prevent a
-						// stale true from eating the next genuine user scroll-up.
-						setTimeout(() => {
-							isProgrammaticScrollRef.current = false;
-						}, 32);
-					}
-				});
-			};
-
-			// Initial scroll on mount/dep change
-			if (shouldAutoScroll()) {
-				scrollToBottom();
-			}
-
-			const observer = new MutationObserver(() => {
-				if (shouldAutoScroll()) {
-					scrollToBottom();
-				}
-			});
-
-			observer.observe(container, {
-				childList: true, // New/removed DOM nodes (new log entries, tool events)
-				subtree: true, // Watch all descendants, not just direct children
-				characterData: true, // Text node mutations (thinking stream text growth)
-			});
-
-			return () => observer.disconnect();
-		}, [autoScrollPaused]);
-
-		// Restore scroll position when component mounts or initialScrollTop changes
-		// Uses requestAnimationFrame to ensure DOM is ready
-		useEffect(() => {
-			// Only restore if we have a saved position and haven't restored yet for this mount
-			if (initialScrollTop !== undefined && initialScrollTop > 0 && !hasRestoredScrollRef.current) {
-				hasRestoredScrollRef.current = true;
-				requestAnimationFrame(() => {
-					if (scrollContainerRef.current) {
-						const { scrollHeight, clientHeight } = scrollContainerRef.current;
-						// Clamp to max scrollable area
-						const maxScroll = Math.max(0, scrollHeight - clientHeight);
-						const targetScroll = Math.min(initialScrollTop, maxScroll);
-						// If the saved position is not at the bottom, pause auto-scroll so the
-						// MutationObserver doesn't immediately yank the view back down (uses the
-						// same 50px bottom threshold as handleScrollInner). Flip the refs first
-						// so the observer's live shouldAutoScroll() sees the pause this frame,
-						// before the state update re-renders.
-						if (targetScroll < maxScroll - 50) {
-							autoScrollPausedRef.current = true;
-							isAtBottomRef.current = false;
-							setAutoScrollPaused(true);
-							setIsAtBottom(false);
-						}
-						scrollContainerRef.current.scrollTop = targetScroll;
-					}
-				});
-			}
-		}, [initialScrollTop]);
-
-		// Reset restore flag when session/tab changes (handled by key prop on TerminalOutput)
-		useEffect(() => {
-			hasRestoredScrollRef.current = false;
-		}, [session.id, activeTabId]);
-
-		// Cleanup throttle timer on unmount
-		useEffect(() => {
-			return () => {
-				if (scrollSaveTimerRef.current) {
-					clearTimeout(scrollSaveTimerRef.current);
-				}
-			};
-		}, []);
-
-		// Helper to find last user command for echo stripping in terminal mode
-		const getLastUserCommand = useCallback(
-			(index: number): string | undefined => {
-				for (let i = index - 1; i >= 0; i--) {
-					if (filteredLogs[i]?.source === 'user') {
-						return filteredLogs[i].text;
-					}
-				}
-				return undefined;
-			},
-			[filteredLogs]
-		);
-
-		// TerminalOutput only handles AI mode; terminal mode renders via TerminalView
-		const isTerminal = false;
-		const isAIMode = true;
-
-		// Memoized prose styles - applied once at container level instead of per-log-item
-		// IMPORTANT: Scoped to .terminal-output to avoid CSS conflicts with other prose containers (e.g., AutoRun panel)
-		const proseStyles = useMemo(
-			() => generateTerminalProseStyles(theme, '.terminal-output'),
-			[theme]
-		);
-
-		const isAutoScrollActive = !autoScrollPaused;
-
-		return (
-			<div
-				ref={terminalOutputRef}
-				tabIndex={0}
-				role="region"
-				aria-label="Terminal output"
-				className="terminal-output flex-1 flex flex-col overflow-hidden transition-colors outline-none relative"
-				style={{
-					backgroundColor: theme.colors.bgMain,
-				}}
-				onKeyDown={(e) => {
-					// Cmd+F to open search
-					if (e.key === 'f' && (e.metaKey || e.ctrlKey) && !outputSearchOpen) {
-						e.preventDefault();
-						setOutputSearchOpen(true);
-						return;
-					}
-					// Escape handling removed - delegated to layer stack for search
-					// When search is not open, Escape should still focus back to input
-					if (e.key === 'Escape' && !outputSearchOpen) {
-						e.preventDefault();
-						e.stopPropagation();
-						// Focus back to text input
-						inputRef.current?.focus();
-						setActiveFocus('main');
-						return;
-					}
-					// Shift+Arrow: jump message-by-message. Skip when the user is typing in
-					// an input/textarea inside the region — those handle their own
-					// arrow-key cursor movement.
-					if (
-						(e.key === 'ArrowUp' || e.key === 'ArrowDown') &&
-						e.shiftKey &&
-						!e.metaKey &&
-						!e.ctrlKey &&
-						!e.altKey &&
-						!isTextInputTarget(e.target)
-					) {
-						const container = scrollContainerRef.current;
-						if (container) {
-							e.preventDefault();
-							jumpToMessageEdge(container, '[data-log-index]', e.key === 'ArrowUp' ? 'up' : 'down');
-						}
-						return;
-					}
-					// Plain Arrow keys: nudge scroll by ~100px (instant, no smooth behavior).
-					if (
-						e.key === 'ArrowUp' &&
-						!e.shiftKey &&
-						!e.metaKey &&
-						!e.ctrlKey &&
-						!e.altKey &&
-						!isTextInputTarget(e.target)
-					) {
-						e.preventDefault();
-						scrollContainerRef.current?.scrollBy({ top: -100 });
-						return;
-					}
-					if (
-						e.key === 'ArrowDown' &&
-						!e.shiftKey &&
-						!e.metaKey &&
-						!e.ctrlKey &&
-						!e.altKey &&
-						!isTextInputTarget(e.target)
-					) {
-						e.preventDefault();
-						scrollContainerRef.current?.scrollBy({ top: 100 });
-						return;
-					}
-					// Option/Alt+Up: page up
-					if (e.key === 'ArrowUp' && e.altKey && !e.metaKey && !e.ctrlKey) {
-						e.preventDefault();
-						const height = terminalOutputRef.current?.clientHeight || 400;
-						scrollContainerRef.current?.scrollBy({ top: -height });
-						return;
-					}
-					// Option/Alt+Down: page down
-					if (e.key === 'ArrowDown' && e.altKey && !e.metaKey && !e.ctrlKey) {
-						e.preventDefault();
-						const height = terminalOutputRef.current?.clientHeight || 400;
-						scrollContainerRef.current?.scrollBy({ top: height });
-						return;
-					}
-					// Cmd+Up to jump to top
-					if (e.key === 'ArrowUp' && (e.metaKey || e.ctrlKey) && !e.altKey) {
-						e.preventDefault();
-						scrollContainerRef.current?.scrollTo({ top: 0 });
-						return;
-					}
-					// Cmd+Down to jump to bottom
-					if (e.key === 'ArrowDown' && (e.metaKey || e.ctrlKey) && !e.altKey) {
-						e.preventDefault();
-						const container = scrollContainerRef.current;
-						if (container) {
-							container.scrollTo({ top: container.scrollHeight });
-						}
-						return;
-					}
-				}}
-			>
-				{/* CSS for Custom Highlight API — paints matches without mutating DOM */}
-				<style>{`
-					::highlight(terminal-search-all) {
-						background-color: ${theme.colors.warning};
-						color: ${theme.mode === 'light' ? '#fff' : '#000'};
-					}
-					::highlight(terminal-search-current) {
-						background-color: ${theme.colors.accent};
-						color: #fff;
-					}
-				`}</style>
-				{/* Output Search */}
-				{outputSearchOpen && (
-					<div
-						className="sticky top-0 z-10 px-3 pt-3 pb-4"
-						style={{ backgroundColor: theme.colors.bgMain }}
-					>
-						<div className="flex items-center gap-2">
-							<div className="relative flex-1">
-								<input
-									type="text"
-									value={outputSearchQuery}
-									onChange={(e) => setOutputSearchQuery(e.target.value)}
-									onKeyDown={(e) => {
-										if (e.key === 'Enter' && !e.shiftKey) {
-											e.preventDefault();
-											goToNextMatch();
-										} else if (e.key === 'Enter' && e.shiftKey) {
-											e.preventDefault();
-											goToPrevMatch();
-										}
-									}}
-									placeholder={
-										outputSearchRegex
-											? 'Regex search... (Enter: next, Shift+Enter: prev)'
-											: 'Search output... (Enter: next, Shift+Enter: prev)'
-									}
-									className="w-full pl-3 pr-14 py-2 rounded border bg-transparent outline-none text-sm"
-									style={{
-										borderColor: regexError ? theme.colors.error : theme.colors.accent,
-										color: theme.colors.textMain,
-										backgroundColor: theme.colors.bgSidebar,
-										fontFamily: outputSearchRegex
-											? 'ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, monospace'
-											: undefined,
-									}}
-									spellCheck={outputSearchRegex ? false : undefined}
-									autoFocus
-								/>
-								<div
-									className="absolute right-2 top-1/2 -translate-y-1/2 px-2 py-0.5 rounded text-xs font-bold pointer-events-none"
-									style={{
-										backgroundColor: theme.colors.bgMain,
-										color: theme.colors.textDim,
-									}}
-								>
-									ESC
-								</div>
-							</div>
-							<button
-								onClick={() => setOutputSearchRegex(!outputSearchRegex)}
-								className="flex items-center justify-center gap-1.5 pl-1 pr-2 rounded border text-xs font-medium whitespace-nowrap transition-colors self-stretch min-w-[7rem]"
-								style={{
-									borderColor: regexError ? theme.colors.error : theme.colors.accent,
-									backgroundColor: theme.colors.accent + '20',
-									color: theme.colors.accent,
-								}}
-								title={outputSearchRegex ? 'Switch to plain-text search' : 'Switch to regex search'}
-							>
-								{/* Pill marker: bg/fg inverted vs. the surrounding button */}
-								<span
-									className="px-1.5 py-0.5 rounded font-mono leading-none"
-									style={{
-										backgroundColor: theme.colors.accent,
-										color: theme.colors.accentForeground,
-									}}
-								>
-									{outputSearchRegex ? '.*' : 'Aa'}
-								</span>
-								<span>{outputSearchRegex ? 'Regex' : 'Plain Text'}</span>
-							</button>
-							{outputSearchQuery.trim() && (
-								<>
-									<span
-										className="text-xs whitespace-nowrap"
-										style={{
-											color: regexError ? theme.colors.error : theme.colors.textDim,
-										}}
-										title={regexError ?? undefined}
-									>
-										{regexError
-											? 'Invalid regex'
-											: totalMatches > 0
-												? `${currentMatchIndex + 1}/${totalMatches}`
-												: 'No matches'}
-									</span>
-									<button
-										onClick={goToPrevMatch}
-										disabled={totalMatches === 0}
-										className="p-1.5 rounded hover:bg-white/10 transition-colors disabled:opacity-30"
-										style={{ color: theme.colors.textDim }}
-										title="Previous match (Shift+Enter)"
-									>
-										<ChevronUp className="w-4 h-4" />
-									</button>
-									<button
-										onClick={goToNextMatch}
-										disabled={totalMatches === 0}
-										className="p-1.5 rounded hover:bg-white/10 transition-colors disabled:opacity-30"
-										style={{ color: theme.colors.textDim }}
-										title="Next match (Enter)"
-									>
-										<ChevronDown className="w-4 h-4" />
-									</button>
-								</>
-							)}
-						</div>
-					</div>
-				)}
-				{/* Prose styles for markdown rendering - injected once at container level for performance */}
-				<style>{proseStyles}</style>
-				{/* Native scroll log list */}
-				{/* overflow-anchor: disabled in AI mode when auto-scroll is off to prevent
-				    browser from automatically keeping viewport pinned to bottom on new content */}
-				<div
-					ref={scrollContainerRef}
-					className="flex-1 overflow-y-auto scrollbar-thin"
-					style={{
-						overflowAnchor: session.inputMode === 'ai' && autoScrollPaused ? 'none' : undefined,
-					}}
-					onScroll={handleScroll}
-				>
-					{/* Log entries */}
-					{filteredLogs.map((log, index) => (
-						<LogItemComponent
-							key={log.id}
-							log={log}
-							index={index}
-							isTerminal={isTerminal}
-							isAIMode={isAIMode}
-							theme={theme}
-							fontFamily={fontFamily}
-							maxOutputLines={maxOutputLines}
-							lastUserCommand={
-								isTerminal && log.source !== 'user' ? getLastUserCommand(index) : undefined
-							}
-							isExpanded={expandedLogs.has(log.id)}
-							onToggleExpanded={toggleExpanded}
-							localFilterQuery={localFilters.get(log.id) || ''}
-							filterMode={filterModes.get(log.id) || { mode: 'include', regex: false }}
-							activeLocalFilter={activeLocalFilter}
-							onToggleLocalFilter={toggleLocalFilter}
-							onSetLocalFilterQuery={setLocalFilterQuery}
-							onSetFilterMode={setFilterModeForLog}
-							onClearLocalFilter={clearLocalFilter}
-							deleteConfirmLogId={deleteConfirmLogId}
-							onDeleteLog={onDeleteLog}
-							onSetDeleteConfirmLogId={setDeleteConfirmLogId}
-							scrollContainerRef={scrollContainerRef}
-							setLightboxImage={setLightboxImage}
-							copyToClipboard={copyToClipboard}
-							ansiConverter={ansiConverter}
-							markdownEditMode={markdownEditMode}
-							onToggleMarkdownEditMode={toggleMarkdownEditMode}
-							onReplayMessage={onReplayMessage}
-							onForkConversation={onForkConversation}
-							sessionId={session.id}
-							onSessionRecover={onSessionRecover}
-							isRecoveringSession={isRecoveringSession}
-							sessionRecoveryError={sessionRecoveryError}
-							fileTree={fileTree}
-							cwd={cwd}
-							projectRoot={projectRoot}
-							onFileClick={onFileClick}
-							sshRemoteId={
-								session.sessionSshRemoteConfig?.enabled
-									? (session.sessionSshRemoteConfig?.remoteId ?? undefined)
-									: undefined
-							}
-							onShowErrorDetails={onShowErrorDetails}
-							onSaveToFile={handleSaveToFile}
-							ghCliAvailable={ghCliAvailable}
-							onPublishGist={onPublishMessageGist}
-							publishedGistUrl={publishedGists[log.id]?.gistUrl}
-							bionifyReadingMode={globalBionifyReadingMode}
-							bionifyIntensity={globalBionifyIntensity}
-							bionifyAlgorithm={globalBionifyAlgorithm}
-							userMessageAlignment={userMessageAlignment}
-							isClaudeCode={session.toolType === 'claude-code'}
-							isAdaptiveMode={getClaudeTokenMode(session) === 'dynamic'}
-						/>
-					))}
-
-					{/* Queued items section - filtered to active tab */}
-					{session.executionQueue && session.executionQueue.length > 0 && (
-						<QueuedItemsList
-							executionQueue={session.executionQueue}
-							theme={theme}
-							onRemoveQueuedItem={onRemoveQueuedItem}
-							onTogglePauseQueuedItem={onTogglePauseQueuedItem}
-							onEditQueuedItem={onEditQueuedItem}
-							onReorderItems={
-								onReorderQueuedItem
-									? (fromIndex, toIndex) =>
-											onReorderQueuedItem(fromIndex, toIndex, activeTabId || undefined)
-									: undefined
-							}
-							onForceSendQueuedItem={onForceSendQueuedItem}
-							forcedParallelEnabled={forcedParallelEnabled}
-							getForceSendContext={getForceSendContext}
-							activeTabId={activeTabId || undefined}
-							onOpenLightbox={setLightboxImage}
-						/>
-					)}
-
-					{/* End ref for scrolling - always rendered so Cmd+Shift+J works even when busy */}
-					<div ref={logsEndRef} />
-				</div>
-
-				{/* Scroll-to-bottom / auto-scroll resume (AI mode only) */}
-				{session.inputMode === 'ai' && filteredLogs.length > 0 && !isAtBottom && (
-					<button
-						onClick={() => {
-							// Jump to bottom and resume auto-scroll
-							setAutoScrollPaused(false);
-							setHasNewMessages(false);
-							setNewMessageCount(0);
-							if (scrollContainerRef.current) {
-								scrollContainerRef.current.scrollTo({
-									top: scrollContainerRef.current.scrollHeight,
-									behavior: 'smooth',
-								});
-							}
-						}}
-						className={`absolute bottom-4 ${userMessageAlignment === 'right' ? 'left-6' : 'right-6'} flex items-center gap-2 px-3 py-2 rounded-full shadow-lg transition-all hover:scale-105 z-20 outline-none`}
-						style={{
-							backgroundColor: isAutoScrollActive
-								? theme.colors.accent
-								: hasNewMessages
-									? theme.colors.accent
-									: theme.colors.bgSidebar,
-							color: isAutoScrollActive
-								? theme.colors.accentForeground
-								: hasNewMessages
-									? theme.colors.accentForeground
-									: theme.colors.textDim,
-							border: `1px solid ${isAutoScrollActive || hasNewMessages ? 'transparent' : theme.colors.border}`,
-							animation:
-								hasNewMessages && !isAutoScrollActive
-									? 'pulse 2s cubic-bezier(0.4, 0, 0.6, 1) infinite'
-									: undefined,
-						}}
-						title={
-							hasNewMessages
-								? 'New messages (click to pin to bottom)'
-								: 'Scroll to bottom (click to pin)'
-						}
-					>
-						<ArrowDown className="w-4 h-4" />
-						{newMessageCount > 0 && !isAutoScrollActive && (
-							<span className="text-xs font-bold">
-								{newMessageCount > 99 ? '99+' : newMessageCount}
-							</span>
-						)}
-					</button>
-				)}
-
-				{/* Copy flash now rendered globally by <CenterFlash /> */}
-
-				{/* Save Markdown Modal */}
-				{saveModalContent !== null && (
-					<SaveMarkdownModal
-						theme={theme}
-						content={saveModalContent}
-						onClose={() => setSaveModalContent(null)}
-						defaultFolder={cwd || session.cwd || ''}
-						isRemoteSession={
-							session.sessionSshRemoteConfig?.enabled && !!session.sessionSshRemoteConfig?.remoteId
-						}
-						sshRemoteId={
-							session.sessionSshRemoteConfig?.enabled
-								? (session.sessionSshRemoteConfig?.remoteId ?? undefined)
-								: undefined
-						}
-						onFileSaved={onFileSaved}
-						onOpenInTab={onOpenInTab}
-					/>
-				)}
-			</div>
-		);
-	})
-);
-
-TerminalOutput.displayName = 'TerminalOutput';
