+++ src/renderer/components/TerminalOutput/index.ts
@@ -1,0 +1,3 @@
+export { TerminalOutput } from './TerminalOutput';
+export { collapseAiResponseLogs } from './utils/collapseAiResponseLogs';
+export type { TerminalOutputProps, LogItemProps, ToolSummary } from './types';
