+++ src/renderer/components/TerminalOutput/utils/logEntryHelpers.ts
@@ -1,0 +1,4 @@
+import type { LogEntry } from '../../../types';
+
+export const isHiddenProgressEntry = (log: LogEntry): boolean =>
+	log.source === 'system' && log.id.startsWith('hidden-progress:');
